third_party/blink/web_tests/html5lib/resources/webkit02.dat

#data
<foo bar=qux/>
#errors
(1,14): expected-doctype-but-got-start-tag
(1,14): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <foo>
|       bar="qux/"

#data
<p id="status"><noscript><strong>A</strong></noscript><span>B</span></p>
#errors
(1,15): expected-doctype-but-got-start-tag
#script-on
#document
| <html>
|   <head>
|   <body>
|     <p>
|       id="status"
|       <noscript>
|         "<strong>A</strong>"
|       <span>
|         "B"

#data
<p id="status"><noscript><strong>A</strong></noscript><span>B</span></p>
#errors
(1,15): expected-doctype-but-got-start-tag
#script-off
#document
| <html>
|   <head>
|   <body>
|     <p>
|       id="status"
|       <noscript>
|         <strong>
|           "A"
|       <span>
|         "B"

#data
<div><sarcasm><div></div></sarcasm></div>
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <sarcasm>
|         <div>

#data
<html><body><img src="" border="0" alt="><div>A</div></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,67): eof-in-attribute-value-double-quote
#new-errors
(1:68) eof-in-tag
#document
| <html>
|   <head>
|   <body>

#data
<table><td></tbody>A
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body
(1,20): foster-parenting-character
(1,20): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "A"
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><td></thead>A
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body
(1,19): XXX-undefined-error
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "A"

#data
<table><td></tfoot>A
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body
(1,19): XXX-undefined-error
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "A"

#data
<table><thead><td></tbody>A
#errors
(1,7): expected-doctype-but-got-start-tag
(1,18): unexpected-cell-in-table-body
(1,26): XXX-undefined-error
(1,27): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>
|             "A"

#data
<legend>test</legend>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <legend>
|       "test"

#data
<table><input>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,14): foster-parenting-start-tag
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <input>
|     <table>

#data
<b><em><dcell><postfield><postfield><postfield><postfield><missing_glyph><missing_glyph><missing_glyph><missing_glyph><hkern><aside></b></em>
#errors
unexpected-b-end-tag
unexpected-em-end-tag
eof-in-aside
#document-fragment
div
#document
| <b>
|   <em>
|     <dcell>
|       <postfield>
|         <postfield>
|           <postfield>
|             <postfield>
|               <missing_glyph>
|                 <missing_glyph>
|                   <missing_glyph>
|                     <missing_glyph>
|                       <hkern>
| <aside>
|   <b>

#data
<b><em><foo><foo><aside></b>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,28): adoption-agency-9
(1,29): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|     <em>
|       <aside>
|         <b>

#data
<b><em><foo><foo><aside></b></em>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,28): adoption-agency-9
(1,33): adoption-agency-9
(1,34): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|     <em>
|     <aside>
|       <em>
|         <b>

#data
<b><em><foo><foo><foo><aside></b>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,33): adoption-agency-9
(1,34): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|             <foo>
|     <aside>
|       <b>

#data
<b><em><foo><foo><foo><aside></b></em>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,33): adoption-agency-9
(1,38): adoption-agency-9
(1,39): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <em>
|         <foo>
|           <foo>
|             <foo>
|     <aside>
|       <b>

#data
<b><em><foo><foo><foo><foo><foo><foo><foo><foo><foo><foo><aside></b></em>
#errors
(1,68): adoption-agency-9
(1,73): adoption-agency-9
(1,74): expected-closing-tag-but-got-eof
#document-fragment
div
#document
| <b>
|   <em>
|     <foo>
|       <foo>
|         <foo>
|           <foo>
|             <foo>
|               <foo>
|                 <foo>
|                   <foo>
|                     <foo>
|                       <foo>
| <aside>
|   <b>

#data
<b><em><foo><foob><foob><foob><foob><fooc><fooc><fooc><fooc><food><aside></b></em>
#errors
(1,77): adoption-agency-9
(1,82): adoption-agency-9
(1,83): expected-closing-tag-but-got-eof
#document-fragment
div
#document
| <b>
|   <em>
|     <foo>
|       <foob>
|         <foob>
|           <foob>
|             <foob>
|               <fooc>
|                 <fooc>
|                   <fooc>
|                     <fooc>
|                       <food>
| <aside>
|   <b>

#data
<option><XH<optgroup></optgroup>
#errors
(1,21): unexpected-start-tag-in-select
(1,32): unexpected-end-tag-in-select
#document-fragment
select
#document
| <option>

#data
<svg><foreignObject><div>foo</div><plaintext></foreignObject></svg><div>bar</div>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,82): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <div>
|           "foo"
|         <plaintext>
|           "</foreignObject></svg><div>bar</div>"

#data
<svg><foreignObject></foreignObject><title></svg>foo
#errors
(1,5): expected-doctype-but-got-start-tag
(1,49): expected-one-end-tag-but-got-another
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|       <svg title>
|     "foo"

#data
</foreignObject><plaintext><div>foo</div>
#errors
(1,16): expected-doctype-but-got-end-tag
(1,16): unexpected-end-tag-before-html
(1,42): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "<div>foo</div>"

#data
<svg xml:base xml:lang xml:space xml:baaah definitionurl>
#errors
no-doctype
eof-in-svg
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       definitionurl=""
|       xml lang=""
|       xml space=""
|       xml:baaah=""
|       xml:base=""

#data
<math definitionurl xlink:title xlink:show>
#errors
no-doctype
eof-in-math
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       definitionURL=""
|       xlink show=""
|       xlink title=""

#data
<math DEFINITIONURL>
#errors
no-doctype
eof-in-math
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       definitionURL=""

#data
<select><hr>
#errors
1:1: ERROR: Expected a doctype token
1:13: ERROR: Premature end of file. Currently open tags: html, body, select.
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <hr>

#data
<select><option><hr>
#errors
1:1: ERROR: Expected a doctype token
1:21: ERROR: Premature end of file. Currently open tags: html, body, select.
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <hr>

#data
<select><optgroup><option><hr>
#errors
1:1: ERROR: Expected a doctype token
1:31: ERROR: Premature end of file. Currently open tags: html, body, select.
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|         <option>
|       <hr>

#data
<select><optgroup><hr>
#errors
1:1: ERROR: Expected a doctype token
1:23: ERROR: Premature end of file. Currently open tags: html, body, select.
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|       <hr>

#data
<select><option><optgroup><hr>
#errors
1:1: ERROR: Expected a doctype token
1:31: ERROR: Premature end of file. Currently open tags: html, body, select.
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <optgroup>
|       <hr>

#data
<table><tr><td><select><hr>
#errors
1:1: ERROR: Expected a doctype token
1:28: ERROR: Premature end of file. Currently open tags: html, body, table, tbody, tr, td, select.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <hr>

#data
<table><tr><td><select><option><hr>
#errors
1:1: ERROR: Expected a doctype token
1:36: ERROR: Premature end of file. Currently open tags: html, body, table, tbody, tr, td, select.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <option>
|               <hr>

#data
<table><tr><td><select><optgroup><option><hr>
#errors
1:1: ERROR: Expected a doctype token
1:46: ERROR: Premature end of file. Currently open tags: html, body, table, tbody, tr, td, select.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <optgroup>
|                 <option>
|               <hr>

#data
<table><tr><td><select><optgroup><hr>
#errors
1:1: ERROR: Expected a doctype token
1:38: ERROR: Premature end of file. Currently open tags: html, body, table, tbody, tr, td, select.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <optgroup>
|               <hr>

#data
<table><tr><td><select><option><optgroup><hr>
#errors
1:1: ERROR: Expected a doctype token
1:46: ERROR: Premature end of file. Currently open tags: html, body, table, tbody, tr, td, select.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <option>
|               <optgroup>
|               <hr>
